Give fixture discovery its own module
The fixture-lockfile gate is about to gain a third operation that has to
cover exactly the same fixture set as the check and refresh modes. Move
discovery to `scripts/fixture_lockfile_discovery.py` so every consumer
resolves one set, relocate `FixtureLockfileError` into the reporting
module (the discovery module cannot raise an error owned by the gate
that imports it without an import cycle), and generalise `GateMode` to
carry the Cargo argv builder and the per-manifest operation instead of
hard-coding locked `cargo metadata`.

Behaviour is unchanged: the check and refresh entry points still run
locked `cargo metadata` per discovered fixture, and the discovery tests
move to `scripts/tests/test_fixture_lockfile_discovery.py`. The gate
script sat at 393 of the 400 lines pylint's `max-module-lines` allows
and its test module at exactly 400, so this split lands before the
feature commit rather than alongside it.

Gates: make check-fmt, make lint-python, make typecheck, and
`pytest scripts/tests` (118 passed) are green.
