+++ scripts/check_fixture_lockfiles.py
@@ -23,7 +23,9 @@ import sys
 import typing as typ
 from pathlib import Path
 
+from fixture_lockfile_discovery import discover_fixture_manifests
 from fixture_lockfile_reporting import (
+    FixtureLockfileError,
     GateMode,
     print_check_summary,
     print_failures,
@@ -35,34 +37,8 @@ from fixture_lockfile_reporting import (
 if typ.TYPE_CHECKING:
     import collections.abc as cabc
 
-#: Directory names the scan never descends into: build output, tool caches and
-#: editor state can all hold generated manifests that are not fixtures.
-EXCLUDED_DIRECTORIES = frozenset({".git", ".vtcode", "target", "node_modules", ".venv"})
-
-#: Manifests that opt out of the workspace but resolve against artefacts staged
-#: under ``target/`` by ``make stage-published-gpui-e2e``. They have no lockfile
-#: to validate until that staging runs, so they stay on their own targets.
-STAGED_FIXTURES = frozenset({"tests/fixtures/published-gpui-e2e"})
-
 CARGO = "cargo"
 METADATA_FORMAT_VERSION = "1"
-
-
-class FixtureLockfileError(RuntimeError):
-    """A fixture lockfile is stale, or no fixture manifest was discovered."""
-
-    @staticmethod
-    def no_manifests_message() -> str:
-        """Return the message for an empty discovery result."""
-        return (
-            "no standalone fixture manifests found; the discovery contract "
-            "expects at least one committed fixture lockfile"
-        )
-
-    @staticmethod
-    def cargo_unavailable_message(cargo: str, manifest: Path, error: OSError) -> str:
-        """Return the message for a Cargo executable that could not run."""
-        return f"cannot run {cargo} for {manifest}: {error}"
 
 
 def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
@@ -75,158 +51,6 @@ def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
     )
     parser.add_argument("--list", action="store_true", help="list the manifests")
     return parser.parse_args(argv)
-
-
-def iter_cargo_manifests(root: Path) -> cabc.Iterator[Path]:
-    """Yield every ``Cargo.toml`` beneath *root*, pruning excluded directories.
-
-    Walking manually (rather than ``Path.rglob``) keeps build output such as
-    ``target/`` and nested scratch copies out of the discovery set.
-
-    Yields
-    ------
-    Path
-        The next manifest beneath *root*, outside the excluded directories.
-    """
-    stack = [root]
-    while stack:
-        directory = stack.pop()
-        try:
-            entries = sorted(directory.iterdir())
-        except OSError:
-            continue
-        for entry in entries:
-            if entry.is_dir() and entry.name not in EXCLUDED_DIRECTORIES:
-                stack.append(entry)
-            elif entry.name == "Cargo.toml" and entry.is_file():
-                yield entry
-
-
-def is_standalone_workspace(manifest: Path) -> bool:
-    """Return whether *manifest* opts out of the root workspace.
-
-    A fixture excludes itself from the enclosing workspace by declaring its own
-    ``[workspace]`` section. The stanza usually carries no members, so the check
-    looks for the section rather than for a member list.
-
-    Parameters
-    ----------
-    manifest : Path
-        The manifest under test.
-
-    Returns
-    -------
-    bool
-        True when the manifest declares its own ``[workspace]`` section.
-    """
-    try:
-        text = manifest.read_text(encoding="utf-8")
-    except OSError:
-        return False
-    return "[workspace]" in text
-
-
-def has_path_dependency(manifest: Path) -> bool:
-    """Return whether *manifest* declares at least one local ``path =`` source.
-
-    Parameters
-    ----------
-    manifest : Path
-        The manifest under test.
-
-    Returns
-    -------
-    bool
-        True when at least one dependency resolves from the local filesystem.
-    """
-    try:
-        text = manifest.read_text(encoding="utf-8")
-    except OSError:
-        return False
-    return "path = " in text
-
-
-def is_staged_fixture(manifest: Path, root: Path) -> bool:
-    """Return whether *manifest* resolves against ``target/`` staged artefacts.
-
-    The published-GPUI end-to-end fixture patches crates.io dependencies onto
-    ``target/published-gpui-e2e/`` package extractions that only exist after
-    ``make stage-published-gpui-e2e``. It therefore has no resolvable graph
-    until staging runs and is validated by the dedicated targets instead.
-
-    Parameters
-    ----------
-    manifest : Path
-        The manifest under test.
-    root : Path
-        The repository root the manifest lives beneath.
-
-    Returns
-    -------
-    bool
-        True when the manifest belongs to a staged fixture directory.
-    """
-    try:
-        return manifest.parent.relative_to(root).as_posix() in STAGED_FIXTURES
-    except ValueError:
-        return False
-
-
-def is_workspace_root(manifest: Path, root: Path) -> bool:
-    """Return whether *manifest* is the root workspace manifest itself.
-
-    The root manifest resolves through the workspace resolver against the root
-    lockfile, which the ordinary workspace build already validates. The fixture
-    gate covers only the crates that opted out of that resolver.
-
-    Parameters
-    ----------
-    manifest : Path
-        The manifest under test.
-    root : Path
-        The repository root the manifest lives beneath.
-
-    Returns
-    -------
-    bool
-        True when *manifest* is the workspace root manifest.
-    """
-    return manifest == root / "Cargo.toml"
-
-
-def discover_fixture_manifests(root: Path) -> list[Path]:
-    """Return every standalone fixture manifest beneath *root*, sorted.
-
-    A manifest qualifies when it opts out of the workspace with ``[workspace]``,
-    uses a local ``path =`` dependency, and commits a sibling ``Cargo.lock``.
-
-    Parameters
-    ----------
-    root : Path
-        The repository root to scan.
-
-    Returns
-    -------
-    list[Path]
-        The sorted authoritative fixture manifests.
-
-    Raises
-    ------
-    FixtureLockfileError
-        No fixture manifest was found, so the gate's contract is broken.
-    """
-    manifests = [
-        path
-        for path in iter_cargo_manifests(root)
-        if is_standalone_workspace(path)
-        and has_path_dependency(path)
-        and (path.parent / "Cargo.lock").is_file()
-        and not is_staged_fixture(path, root)
-        and not is_workspace_root(path, root)
-    ]
-    if not manifests:
-        raise FixtureLockfileError(FixtureLockfileError.no_manifests_message())
-    return sorted(manifests)
 
 
 def cargo_metadata_command(manifest: Path) -> list[str]:
@@ -306,26 +130,27 @@ def refresh_lockfile(manifest: Path) -> subprocess.CompletedProcess[str]:
 
 
 def collect_fixture_results(
-    root: Path,
     manifests: list[Path],
+    operation: cabc.Callable[[Path], subprocess.CompletedProcess[str]],
     prepare: cabc.Callable[[Path], object] | None = None,
 ) -> list[tuple[Path, subprocess.CompletedProcess[str]]]:
-    """Return the failing locked-metadata results for *manifests*.
+    """Return the failing results of *operation* over *manifests*.
 
     Each manifest is optionally prepared (refresh mode regenerates its lockfile
-    first), then validated with locked ``cargo metadata``; the loop never
-    stops at the first failure so one gate run reports every stale fixture.
+    first), then run through *operation*, the locked ``cargo metadata``
+    validation both modes share. The loop never stops at the first failure so
+    one gate run reports every fixture.
 
     Returns
     -------
     list[tuple[Path, subprocess.CompletedProcess[str]]]
-        Every ``(manifest, result)`` pair whose lockfile failed to resolve.
+        Every ``(manifest, result)`` pair whose operation failed.
     """
     failures: list[tuple[Path, subprocess.CompletedProcess[str]]] = []
     for manifest in manifests:
         if prepare is not None:
             prepare(manifest)
-        result = run_cargo_metadata(manifest)
+        result = operation(manifest)
         if result.returncode != 0:
             failures.append((manifest, result))
     return failures
@@ -337,11 +162,11 @@ def _report_fixture_operation(
     mode: GateMode,
 ) -> int:
     """Collect, report, and summarize one gate operation over *manifests*."""
-    results = collect_fixture_results(root, manifests, prepare=mode.prepare)
+    results = collect_fixture_results(manifests, mode.operation, prepare=mode.prepare)
     failures = [
         mode.failure_message(
             manifest.relative_to(root),
-            cargo_metadata_command(manifest),
+            mode.command(manifest),
             result.stdout,
             result.stderr,
         )
@@ -358,7 +183,12 @@ def check_fixtures(root: Path, manifests: list[Path]) -> int:
     return _report_fixture_operation(
         root,
         manifests,
-        GateMode(stale_failure_message, print_check_summary),
+        GateMode(
+            stale_failure_message,
+            print_check_summary,
+            cargo_metadata_command,
+            run_cargo_metadata,
+        ),
     )
 
 
@@ -367,7 +197,13 @@ def refresh_fixtures(root: Path, manifests: list[Path]) -> int:
     return _report_fixture_operation(
         root,
         manifests,
-        GateMode(refresh_failure_message, print_refresh_summary, refresh_lockfile),
+        GateMode(
+            refresh_failure_message,
+            print_refresh_summary,
+            cargo_metadata_command,
+            run_cargo_metadata,
+            refresh_lockfile,
+        ),
     )
